This branch copies py files, resource files and Arcus.pyd from the 6.5.4 application distributed by QIDI, it works for now, but there are errors.
I think it's probably to save space on Google Drive, but if possible, please distribute not only the latest version, but also previous versions of the software.

+++ plugins/GCodeWriter/utils.py
@@ -112,7 +112,8 @@ def add_screenshot(img, width, height, img_type):
         for i in range(len(outputdata)):
             if (outputdata[i] != 0):
                 astr += chr(outputdata[i])
-        os.remove(os.path.abspath("")+"\\cut_image_.png")
+        if os.path.exists(os.path.abspath("")+"\\cut_image_.png"):
+            os.remove(os.path.abspath("")+"\\cut_image_.png")
     except Exception as e:
         Logger.log("d", "Exception == " + str(e))
